Split gossip interface from inmem
